+++ packages/react/CHANGELOG.md
@@ -1,12 +1,12 @@
-# @stella-ui/react
+# @stella-ds/react
 
 ## 0.2.0
 
 ### Minor Changes
 
-- 747cbf6: Initial release of @stella-ui/react and @stella-ui/theme
+- 747cbf6: Initial release of @stella-ds/react and @stella-ds/theme
 
 ### Patch Changes
 
 - Updated dependencies [747cbf6]
-  - @stella-ui/theme@0.2.0
+  - @stella-ds/theme@0.2.0